Move Qt primitives out of "common" to fix non-Qt builds (#2299)
* Move Qt primitives out of "common" to fix non-Qt builds

* "include" path fix for Qt primitive move

* Conditionally include Qt in lang install

* Fix a CMake path

* "include" fix for Qt breaking builds
