+++ editors/sc-ide/primitives/localsocket_utils.cpp
