+++ lang/CMakeLists.txt
@@ -34,6 +34,12 @@ find_package(Readline 5.0)
 
 find_package(Sndfile)
 
+if (SC_QT OR SC_IDE)
+	set(LOCALSOCK_UTIL_FILE ${CMAKE_SOURCE_DIR}/editors/sc-ide/primitives/localsocket_utils.cpp)
+else ()
+	set(LOCALSOCK_UTIL_FILE)
+endif ()
+
 set(sclang_sources
 	LangPrimSource/SC_HID_api.cpp
 	LangPrimSource/SC_Wii.cpp
@@ -83,7 +89,7 @@ set(sclang_sources
 	${CMAKE_SOURCE_DIR}/common/SC_StringBuffer.cpp
 	${CMAKE_SOURCE_DIR}/common/SC_StringParser.cpp
 	${CMAKE_SOURCE_DIR}/common/SC_TextUtils.cpp
-	${CMAKE_SOURCE_DIR}/common/localsocket_utils.cpp
+	${LOCALSOCK_UTIL_FILE}
 
 	${CMAKE_SOURCE_DIR}/common/sc_popen.cpp
 )
